added mapper

+++ beauty-studio-service/src/main/java/com/company/service/impl/CategoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.company.service.impl;
 
 import com.company.entity.CategoryEntity;
 import com.company.repository.CategoryRepository;
+import com.company.repository.ProcedureRepository;
 import com.company.service.CategoryService;
 import com.company.service.dto.CategoryDto;
 import com.company.service.exception.HttpStatusCode;
@@ -17,8 +18,10 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
+
     private final CategoryMapper categoryMapper;
     private final CategoryRepository categoryRepository;
+    private final ProcedureRepository procedureRepository;
 
     @Override
     @Transactional
@@ -28,17 +31,17 @@ public class CategoryServiceImpl implements CategoryService {
             throw new ServiceException(HttpStatusCode.CONFLICT, "Category with this name = " + categoryDto.getName() + " already exist!");
         }
 
-        CategoryEntity categoryEntity = categoryMapper.convertToEntity(categoryDto);
+        CategoryEntity categoryEntity = categoryMapper.toEntity(categoryDto);
 
         CategoryEntity saved = categoryRepository.save(categoryEntity);
 
-        return categoryMapper.convertToDto(saved);
+        return categoryMapper.toDto(saved);
     }
 
     @Override
     public CategoryDto findById(Long categoryId) {
         CategoryEntity categoryEntity = getCategoryEntity(categoryId);
-        return categoryMapper.convertToDto(categoryEntity);
+        return categoryMapper.toDto(categoryEntity);
     }
 
     private CategoryEntity getCategoryEntity(Long categoryId) {
@@ -50,12 +53,17 @@ public class CategoryServiceImpl implements CategoryService {
     public List<CategoryDto> findAll() {
         List<CategoryEntity> categories = categoryRepository.findAll();
 
-        return categoryMapper.convertListToDto(categories);
+        return categoryMapper.toDtoList(categories);
     }
 
     @Override
     public void delete(Long categoryId) {
+
         CategoryEntity categoryEntity = getCategoryEntity(categoryId);
+
+        if (procedureRepository.existsProcedureEntityByCategoryId(categoryEntity.getId())) {
+            throw new ServiceException(HttpStatusCode.CONFLICT, "Category contains procedures, delete them firstly");
+        }
 
         categoryRepository.delete(categoryEntity);
     }
